Extract weather widget
